+++ src/test/suite/utils/static/directiveNames.ts
@@ -14,9 +14,7 @@ export const DIRECTIVES = [
     '@for @else',
     '@dump',
     '@end',
-    '@slot',
-    '@slot(name)',
 ]
 
 export const INSIDE_LOOP_DIRECTIVES = ['@break', '@continue', '@continueIf', '@breakIf']
-// export const INSIDE_COMP_DIRECTIVES = ['@slot', '@slot(name)']
+export const INSIDE_COMP_DIRECTIVES = ['@slot', '@slot(name)']